+++ relay-server/src/integrations/mod.rs
@@ -42,6 +42,7 @@ define_integrations!(
     "application/vnd.sentry.integration.otel.logs+protobuf" => Integration::Logs(LogsIntegration::OtelV1 { format: OtelFormat::Protobuf }),
     "application/vnd.sentry.integration.otel.spans+json" => Integration::Spans(SpansIntegration::OtelV1 { format: OtelFormat::Json }),
     "application/vnd.sentry.integration.otel.spans+protobuf" => Integration::Spans(SpansIntegration::OtelV1 { format: OtelFormat::Protobuf }),
+    "application/vnd.sentry.integration.vercel.logs+json" => Integration::Logs(LogsIntegration::VercelDrainLog),
 );
 
 /// An exhaustive list of all integrations supported by Relay.
@@ -74,6 +75,10 @@ pub enum LogsIntegration {
     ///
     /// Supports OTeL's [`LogsData`](opentelemetry_proto::tonic::logs::v1::LogsData).
     OtelV1 { format: OtelFormat },
+    /// The Vercel Log Drain integration.
+    ///
+    /// Supports the [`relay_ourlogs::VercelLog`] format.
+    VercelDrainLog,
 }
 
 /// All span integrations supported by Relay.